moved classes and tests to other packages
