+++ src/main/kotlin/com/framework/test/application/config/factory/ApplicationConfigReader.kt
@@ -1,4 +1,4 @@
-package com.framework.test.config.provider.applicationConfig.factory
+package com.framework.test.application.config.factory
 
 import com.framework.test.model.ApplicationConfig
