+++ src/main/kotlin/com/framework/test/application/config/factory/YamlApplicationConfigReaderFactory.kt
@@ -1,20 +1,22 @@
-package com.framework.test.config.provider.applicationConfig.factory
+package com.framework.test.application.config.factory
 
-import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
 import com.fasterxml.jackson.module.kotlin.readValue
 import com.framework.test.constants.FileNames
 import com.framework.test.file.reader.FileProvider
 import com.framework.test.model.ApplicationConfig
 import java.io.File
 
-class JsonApplicationConfigReaderFactory(
+class YamlApplicationConfigReaderFactory(
   private val fileProvider: FileProvider = FileProvider()
 ) : ApplicationConfigReader {
-  override val configFilePath: String = "authData/${FileNames.BASIC_AUTH_FILE_NAME.value}.json"
+  override val configFilePath: String = "authData/${FileNames.BASIC_AUTH_FILE_NAME.value}.yaml"
   private val file: File by lazy { fileProvider.getFileFromPath(configFilePath) }
 
   override fun getApplicationConfigFromFile(): ApplicationConfig {
-    val mapper = jacksonObjectMapper()
+    val mapper = ObjectMapper(YAMLFactory()).findAndRegisterModules() // worksWithJsonToo
     return mapper.readValue(file)
   }
 }
+
